Day 13. Continue hot reloading implementation
The past few days have consisted in a slow dance of opening the
 source files, poking a bit around, making refactorings here and
 there, reading some documentation, and closing it all. This
 resulted however in touching almost all of my codebase to the
 following chages:

* I started on a journey of incorporating my yak_defines.h inside
the project. For now, I changed the includes to types.h which then
redirects to yak_defines. The idea is that, once, I'm happy with the
state of yak_defines, I'll simply copy the code from there into the
types.h.
* sparrow.c: compressing _DllMainCRTStartup even further. Not
sure I can get rid of it entirely without linking CRT.
* win32_sparrow.c: continue work on live code reloading. My
current approach is to rely on existance of the lock file, but a
concrete solution still hasn't presented itself (and I'm not entirely
happy with the one proposed by Casey). In the meantime, I've
continued definition of the win32_module structure. I hope to find
a solution to this in the next commit.