Day 17. Fixing pixel grid orientation
The efforts to improve the console pixel grid continue. Today, the
biggest implementation was that of the `GetPixel` function. We
can now properly (and quickly?) load any pixel with the specified
coordinates. The function itself its pretty simple, if you ignore all
the assertions (that fired oh so many times during its implementation).

    local struct pixel*
    GetPixel(struct pixel* Pixels, v2u Coords)
    {
        u32 Y = CONSOLE_WIDTH * Coords.y;
        u32 Pos = Y + Coords.x;

        struct pixel* Result = Pixels + Pos;

        return (Result);
    }

Even this is a bit of an overkill but I wanted to be super explicit,
should we need to revisit this function later down the line. For
now, the plan is to use the new function to assist in implementing
the first input system.

One big issue that was fixed was the coordinate system
orientation of the pixel grid. Before today's work, the origin of the
console grid was in the bottom-left corner of the buffer; now it's
in the top-left. Hopefully this will allow easier iteration moving
forward.