Day 20. Debugging pixel size
Today was a bit of an exercise in futility. In my rush to implement a
rendering system I actually introudced a whole bunch of bugs that
need to be addressed.

`win32_sparrow.c`: Removed fixed buffer size at initialization
`sparrow_console_render.c`: Identified an issue where sometimes a
pixel is too big, and sometimes is too small. A proper stretching
mechanism will need to be implemented if we are to solve this issue.

For the most part I have reverted my changes to the start of the
day. I'll need to revisit console rendering code at a later stage.