Day 21. Revert DrawPixel changes
My attempts in removing scaling artifacts highlighted that I had a
very vague understanding of the problem at hand. By now I
mostly reverted the changes of the last few days except I
streamlined the DrawPixel call a bit. I will need to revisit this part
of code in the future; for now, as exploration code, it's more than
enough.