Day 31. Writing more vector helpers
I added Add, Subtract, Multiply and Divide functions to the
different vector scenarios, as well as made extensions/shortcuts
for C++. I then plugged it all in the viewer to test it out.

One thing that immediately jumped to my attention that I have
swapped top and bottom. I mean, it kind of was a point
beforehand, but I'm starting to think that I need to account for the
origin position during the console definition. Will try to tackle it
next time.

Also yay, this project is officially one "month" old! Or about half a
year, depends on your point of view.