## Day 38. Debugging Diffuse mapping
I started the day by chasing the yesterday's graphical bug. I soon
confirmed that it's indeed called "[Shadow Acne]
(https://digitalrune.github.io/DigitalRune-Documentation/html/3f4d959e-9c98-4a97-8d85-7a73c26145d7.htm)"
 and can be pretty nasty. It seems that here finally we come to the
 point where the double precision arithmetic prevails over the
 floating point.

![Shadow acne bug in the glorious 1080p](media/day38/shadow_acne_1.png)

At the end, I cranked up the epsilon to 0.1f and moved on.

![Shadow acne with 0.01f epsilon](media/day38/shadow_acne_2.png)

Another interesting bug that I found is however more mysterious.
From time to time, the rendering freaks out completely and shows
an outworldly bogus output:

![Bogus output](media/day38/rendering_bug.png)

I suspect that this has to do with the way random number
generator operates, or something. Honestly, I have no idea about
what's going on, and why is it sporadic. In any case, this project so
far has gifted us a truly marvellous collection of graphical bugs.

Going back to shadow acne, after applying a different diffuse
method ("True Lambertian Reflection"), I realized that it's still
present there, maybe even clearer:

![Shadow acne in Lambertian reflection](media/day38/shadow_acne_3.png)

Everything became clear once I removed the diffusing at all:

![True reflection](media/day38/shadow_acne_4.png)

It's not about the diffuse at all! We were somehow receiving data
with a terrible precision. It wasn't sure if it was about float vs
double anymore, but it kind of was the biggest offender so I dug
deeper to investigate the bug.

Right now we are getting pure reflections and shadows by polling
multiple levels of recursion depth. If I gradually increase it, we get
the following progression:

![Effect of the recursion depth increase]
(media/day38/depth-evolution.gif)

What we see here that yes, increasing depth does improve the
visual fidelity of the output, but there's more to it. There's no
difference between recursion depth 20 and 1000.

Next, what we see from this is, that the sky is reflected accurately,
it's the normals of the sphere which are flat-shaded until they hit
the next "tier". This may be explained in the following manner:

- We use truncation somewhere and information is lost
- The floating point precision is insufficient
- ???

I tried to visualize different information that I receive with the first
hit (depth = 1). Seems in order, and I understood that this trail was
fine. I had only one avenue remaining: the diffuse algorithm itself.

![Hit record](media/day38/hit-record.png)

So, at the end, I realized that my random number generator was
returning garbage. It _was_ a truncation error, and by adding a
couple assertions the problem was quickly identified and resolved.
It was a fun journey.

![Perfect Diffuse](media/day38/perfect_diffuse.png)

On a completely unrelated note, I'll have to move to compute
shaders on my graphics card, if I'm to continue this journey.
Already now I'm only rendering in release mode, unless I want to
wait for 15 seconds+ to draw a single frame.